code re-written, finishing hibernate

+++ service/src/main/java/com/gmail/kurmazpavel/CatalogService.java
@@ -3,6 +3,7 @@ package com.gmail.kurmazpavel;
 import com.gmail.kurmazpavel.beans.dto.CatalogDTO;
 import com.gmail.kurmazpavel.beans.dto.DiscountDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CatalogService {
@@ -11,7 +12,7 @@ public interface CatalogService {
     CatalogDTO update(CatalogDTO catalogDTO);
     CatalogDTO delete (CatalogDTO catalogDTO);
     List<CatalogDTO> getAll();
-    Long count (Double min, Double max);
-    List<CatalogDTO> getByPrice(Double min, Double max);
+    Long count (BigDecimal min, BigDecimal max);
+    List<CatalogDTO> getByPrice(BigDecimal min, BigDecimal max);
     List<DiscountDTO> getDiscounts(Long id);
 }
